Refactor, no functional change.
